added exception handling when no result - 404 fixed bugs in HolidayRestEndpoint controller

+++ src/main/java/si/arctur/work/calendar/service/HolidayService.java
@@ -1,5 +1,6 @@
 package si.arctur.work.calendar.service;
 
+import org.apache.commons.lang3.Validate;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,6 +14,7 @@ import si.arctur.work.calendar.dao.repository.HolidayRepository;
 import si.arctur.work.calendar.model.HolidayDTO;
 import java.time.LocalDate;
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Component
@@ -41,7 +43,10 @@ public class HolidayService {
     public List<HolidayDTO> getHolidaysForCalendar(Long calendarId) {
         LOG.info("START - getHolidaysForCalendar(calendarId={})", calendarId);
 
-        Assert.isNull(calendarId, "calendarId attribute must not be null!");
+        if(Objects.isNull(calendarId)) {
+            LOG.error("calendarId attribute must not be null!");
+            throw new IllegalArgumentException("calendarId attribute must not be null!");
+        }
 
         return holidayRepository.getHolidayEntitiesByWorkCalendars(new WorkCalendarEntity(calendarId)).stream()
                 .map(holiday -> holidayConverter.convert(holiday))
@@ -51,33 +56,53 @@ public class HolidayService {
     public HolidayDTO getHoliday(Long id) {
         LOG.info("START - getHoliday(id={})", id);
 
-        Assert.isNull(id, "id attribute must not be null!");
+        if(Objects.isNull(id)) {
+            LOG.error("id attribute must not be null!");
+            throw new IllegalArgumentException("id attribute must not be null!");
+        }
 
         return holidayConverter.convert(holidayRepository.getHolidayEntityById(id));
     }
 
+    /**
+     * Add new holiday without adding reference to work calendar
+     * @param holidayDTO
+     * @return
+     */
     public HolidayDTO addHoliday(HolidayDTO holidayDTO) {
         LOG.info("START - addHoliday(holidayDTO={})", holidayDTO);
 
-        Assert.isNull(holidayDTO, "HolidayDTO object must not be null!");
+        if(Objects.isNull(holidayDTO)) {
+            LOG.error("holidayDTO attribute must not be null!");
+            throw new IllegalArgumentException("holidayDTO attribute must not be null!");
+        }
 
         HolidayEntity holidayEntity = new HolidayEntity();
         holidayEntity.setDate(holidayDTO.getDate());
         holidayEntity.setName(holidayDTO.getName());
         holidayEntity.setWorkFree(holidayDTO.getWorkFree());
 
-        //TODO: check if all ok
-        holidayEntity.getWorkCalendars().addAll(holidayDTO.getWorkCalendars().stream()
-                .map(workCalendarDTO -> new WorkCalendarEntity(workCalendarDTO.getId()))
-                .collect(Collectors.toList()));
+        //add holiday to all selected calendars
+//        holidayEntity.getWorkCalendars().addAll(holidayDTO.getWorkCalendars().stream()
+//                .map(workCalendarDTO -> new WorkCalendarEntity(workCalendarDTO.getId()))
+//                .collect(Collectors.toList()));
 
         return holidayConverter.convert(holidayRepository.save(holidayEntity));
     }
 
+    /**
+     * Update selected holiday, without changing reference to work calendar
+     * @param holidayDTO
+     * @return
+     */
     public HolidayDTO updateHoliday(HolidayDTO holidayDTO) {
         LOG.info("START - updateHoliday(holidayDTO={})", holidayDTO);
 
-        Assert.isNull(holidayDTO, "HolidayDTO object must not be null!");
+        if(Objects.isNull(holidayDTO)) {
+            LOG.error("HolidayDTO attribute must not be null!");
+            throw new IllegalArgumentException("HolidayDTO attribute must not be null!");
+        }
+//        Assert.notNull(holidayDTO, "HolidayDTO object must not be null!");
 
         HolidayEntity holidayEntity = holidayRepository.getHolidayEntityById(holidayDTO.getId());
 
@@ -85,9 +110,9 @@ public class HolidayService {
         holidayEntity.setName(holidayDTO.getName());
         holidayEntity.setDate(holidayDTO.getDate());
 
-        holidayEntity.setWorkCalendars(holidayDTO.getWorkCalendars().stream()
-                .map(workCalendarDTO -> new WorkCalendarEntity(workCalendarDTO.getId()))
-                .collect(Collectors.toSet()));
+//        holidayEntity.setWorkCalendars(holidayDTO.getWorkCalendars().stream()
+//                .map(workCalendarDTO -> new WorkCalendarEntity(workCalendarDTO.getId()))
+//                .collect(Collectors.toList()));
 
         return holidayConverter.convert(holidayRepository.save(holidayEntity));
     }
@@ -95,7 +120,10 @@ public class HolidayService {
     public void deleteHoliday(Long id) {
         LOG.info("START - deleteHoliday(id={})", id);
 
-        Assert.isNull(id, "id attribute must not be null!");
+        if(Objects.isNull(id)) {
+            LOG.error("id attribute must not be null!");
+            throw new IllegalArgumentException("id attribute must not be null!");
+        }
 
         holidayRepository.delete(new HolidayEntity(id));
     }
